+++ src/pas/plugins/kimug/setuphandlers/__init__.py
@@ -1,4 +1,6 @@
 from pas.plugins.kimug.plugin import KimugPlugin
+
+# from pas.plugins.kimug.utils import remove_authentic_plugin
 from pas.plugins.kimug.utils import clean_authentic_users
 from pas.plugins.kimug.utils import get_keycloak_users
 from pas.plugins.kimug.utils import migrate_plone_user_id_to_keycloak_user_id
@@ -58,6 +60,6 @@ def post_install(context):
                 kc_users,
             )
             clean_authentic_users()
-            remove_authentic_plugin()
+            # remove_authentic_plugin()
         else:
             logger.error(f"Keycloak realm '{keycloak_realm}' does not exist.")