+++ CHANGES.md
@@ -1,7 +1,20 @@
-## 1.1.1 (unreleased)
+## 1.1.3 (unreleased)
 
 
 - Nothing changed yet.
+
+
+## 1.1.2 (2025-08-27)
+
+
+- Add fogrot local roles on migration to Keycloak.
+  [bsuttor & rdubois]
+
+## 1.1.1 (2025-08-26)
+
+
+- Migrate users form Authentic to Keycloal OIDC plugin.
+  [bsuttor]
 
 
 ## 1.1.0 (2025-07-10)